requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
redis==5.2.0
hiredis==3.0.0
solana==0.36.0
solders==0.23.0
base58==2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
cryptography==44.0.0
pydantic[email]==2.10.0
pydantic-settings==2.6.0
argon2-cffi==23.1.0
PyNaCl==1.5.0
boto3==1.35.0
botocore==1.35.0
slowapi==0.1.9
limits==3.13.0
python-dateutil==2.9.0
sentry-sdk[fastapi]==2.19.0
structlog==24.4.0
python-json-logger==2.0.7
celery==5.4.0
flower==2.0.1
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.0
faker==33.0.0
factory-boy==3.3.1
email-validator==2.2.0
orjson==3.10.0
gunicorn==23.0.0
python-decouple==3.8
